add ConsoleTest project
