+++ Paging.sln
@@ -12,6 +12,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.Tests", "tests\Pagin
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.Benchmarks", "benchmarks\Paging.Benchmarks\Paging.Benchmarks.csproj", "{D797B4BE-4C73-42A2-908C-208CC1244967}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.ConsoleTests", "tests\Paging.ConsoleTests\Paging.ConsoleTests.csproj", "{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,7 @@ Global
 		{C85597B6-A4C7-4E0C-854B-66FD5F27F33B} = {D7D757C5-E956-428C-80F7-09F391D9E9DE}
 		{A77C3D96-BA31-4032-B7E6-008EADFD9280} = {DCA445CD-81D1-461E-8066-923FB17477EB}
 		{D797B4BE-4C73-42A2-908C-208CC1244967} = {D7066DB3-706B-4D98-A0B1-D2553BCB61A2}
+		{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC} = {DCA445CD-81D1-461E-8066-923FB17477EB}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{C85597B6-A4C7-4E0C-854B-66FD5F27F33B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -35,5 +38,9 @@ Global
 		{D797B4BE-4C73-42A2-908C-208CC1244967}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D797B4BE-4C73-42A2-908C-208CC1244967}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D797B4BE-4C73-42A2-908C-208CC1244967}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
