+++ benchmarks/Paging.Benchmarks/Paging.Benchmarks.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net8.0<!--;net7.0;netstandard2.0;netstandard2.1--></TargetFrameworks>
+		<TargetFramework>net8.0</TargetFramework>
 		<LangVersion>default</LangVersion>
 		<SignAssembly>true</SignAssembly>
