+++ src/Paging/Paging.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net8.0<!--;net7.0;netstandard2.0;netstandard2.1--></TargetFrameworks>
+		<TargetFrameworks>net8.0;net7.0;net6.0;netstandard1.3;netstandard2.0;netstandard2.1;net48;net47;net45</TargetFrameworks>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 		<LangVersion>default</LangVersion>
@@ -30,6 +30,14 @@
 	<PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
 	  <DocumentationFile>bin\Release\Paging.xml</DocumentationFile>
 	</PropertyGroup>
+
+	<PropertyGroup Condition=" '$(TargetFramework)' != 'net8.0' AND '$(TargetFramework)' != 'net7.0' AND '$(TargetFramework)' != 'net6.0'">
+		<DefineConstants>DOES_NOT_SUPPORT_JSON</DefineConstants>
+	</PropertyGroup>
+
+	<ItemGroup Condition=" '$(TargetFramework)' != 'net8.0' AND '$(TargetFramework)' != 'net7.0' AND '$(TargetFramework)' != 'net6.0'">
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+	</ItemGroup>
 
 	<ItemGroup>
 		<None Include="README.md" Pack="true" PackagePath="\"/>
