+++ tests/Paging.ConsoleTests/Program.cs
@@ -1,0 +1,5 @@
+﻿// See https://aka.ms/new-console-template for more information
+
+using Paging.ConsoleTests.Tests;
+
+JsonSerializationTests.Test();
